// start/frontend/src/app/model/nekretnina.model.ts
export class Nekretnina{
    id:Number;
    naziv: String;
    grad: String;
    opstina: String;
    ulica: String;
    tip: String;
    brojSpratova: String;
    sprat: Number;
    kvadratura: Number;
    brojSoba: Number;
    jeNamestena: Boolean;
    galerija: [String];
    usluga: String;
    cena: Number;
    vlasnik:String;    
    jePromovisana: String;
    stanje: String;
}